escola/requirements.txt
Django==2.2.9
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1
